add test, depend on spark-testkit
