+++ build.sbt
@@ -7,10 +7,11 @@ resolvers += "LocationTech GeoTrellis Snapshots" at "https://repo.locationtech.o
 
 def geotrellis(module: String) = "org.locationtech.geotrellis" %% s"geotrellis-$module" % "1.0.0-SNAPSHOT"
 
-def spark(module: String) = "org.apache.spark" %% s"spark-$module" % "2.0.2"
+def spark(module: String) = "org.apache.spark" %% s"spark-$module" % "2.1.0"
 
 libraryDependencies ++= Seq(
   geotrellis("spark"),
+  geotrellis("spark-testkit") % Test,
   spark("core"),
   spark("sql"),
   "com.databricks" %% "spark-avro" % "3.1.0",
